Solving step 27

+++ Liquid.cpp
@@ -11,7 +11,7 @@
 
 using namespace std;
 
-Liquid::Liquid(string marca, int volume, double valor) : Food(valor) {
+Liquid::Liquid(string marca, double volume, double valor) : Food(valor) {
     this->volume = volume;
     this->marca = marca;
 }
